update README.txt
This is a small-scale web scraper I made in resposne to a coding challenge in October of 2018. 

There are two files here: the first .py file contains the spider's code, and the second is a text file that contaisn the spiders code w/ my comments. These are seperate because my IDE was, for whatever reason, unable to compile my code when accompanied by all my comments. I was given an error relating to indentation, specifically one saying the compiler could not figure out how to decipher pieces of code nested and indented in my for loops due to the placement of my comments. It's an odd bug, so for the sake of relieving potential headaches and frusteration I just seperated the files altogether.

Enjoy!